First pass at a purecap bbl.
I started by trying to have a hybrid bbl that just saved capregs instead
of gprs, but the machine/ bits really assume registers are uintptr_t's
all over the place (better than assuming longs!).  I then thought about
saving capregs as well as gprs, but that would mean that any operations
that modified a GPR would have to also modify the capability so that
I could always restore from capregs.  It seemed simplest to instead just
go for purecap.

While this in theory updates all of the exception entry/exit points
as well as replacing csr references with scrs for redirected traps, it
does not yet compile as a purecap program.