Bounce unhandled misaligned accesses as-is, not as access faults for insn
We hit this case for load/store capability instructions (and in theory
all the explicit loads/stores). Reporting it as an illegal instruction
(due to not being able to decode it) would be one thing (though not what
OpenSBI does), but changing it to an access fault whilst also using the
instruction bits as tval is completely insane and causes the CheriBSD
kernel to (rightly) turn such a trap into a SIGSEGV for the process.
Kernels may not be prepared to get an alignment fault given M-mode is
meant to handle it and pretend they don't exist, but it is the right
thing to do.