Fix BSS zeroing for multicore
When I upstreamed support for this (which, incidentally, is guarded by
an off-by-default configure option) I fixed it to be in the right place
for multicore, i.e. to only be done once by the boot hart, but never
fixed our fork.

Whilst here, slightly reduce the diff to that upstream version so it's
less confusing if/when we merge.