Avoid using soon-to-be mode-dependent instructions
The __trap_from_machine_mode change is slightly inefficient but will go
away once we can rely on JAL being CJAL in capability mode. Note the
existing code was already broken; whilst it ended up working fine in the
PCC-relative ABI for the no return bad_trap case, it would take a tag
fault on the return path if it ever hit one of the non-fatal cases
(related to trapping whilst emulating an instruction, though I believe
this only happens if there is a second hart modifying the page tables,
or if the hart has not executed an sfence.vma since modifying the page
tables, the old versions remain cached in the TLB for the original
attempted execution and then get evicted at some point during
emulation, making these cases exceedingly rare to hit).

The fp_asm.S changes are technically unnecessary in the PCC-relative
ABI given this is a statically-linked binary, but will no longer
assemble once JALR is made mode-dependent.